Phone Orientation
In this project, we create an application with the objective to provide a personal feed for the user to keep track of any recent purchases or products or whatever the user owns. The main components of this project include Authentication Flow, Adding new products, Edit and Verify Profile options, Search and Display of the products added by the user. The front end is completely done in Android Studio whereas the backend is implemented using Firebase.