Location Tracking
In this application, we will be using GPS and accelerometer to determine the current location of a device. Initially, we display the current location and display it on the map and then later use accelerometer to calculate approximate location using the initial position, accelerometer, time and current location in the map. The initial locations are used to calculate the future locations and accelerometer data is used to improve accuracy. To save battery power, instead of going for the highest accuracy we can trade-off between power and accuracy and use other techniques as mentioned above to obtain better location accuracy.