Authentication App using AWS
In this application, we create a sample authentication flow for the user to register, confirm their registration, login into an existing account and reset their password if forgotten. Initially, the login page will be displayed on default if the user opens the application. The user can directly login into the application with their credentials if they have an account. If they don’t have an account, there is an option to open the registration page which allow the user to register with the application followed by a confirmation page for the registration. If the user forgot their password, it can also be reset. In the backend, we are using Amazon Web Services namely Lambda, Cognito and API Gateway and boto3 package functions for creating new users, accessing them, and many more.